refactor(frontend): 💡 separate selectors into separate file
