+++ src/features/sheets/store/index.ts
@@ -9,3 +9,5 @@ export * from "./newSheetPageState";
 export * from "./newSheetPageState";
 export * from "./sheetPageState";
 export * from "./useInventoryStore";
+export * from "./inventorySelectors";
+export * from "./inventoryActions";
