+++ src/features/sheets/store/useInventoryStore.ts
@@ -1,6 +1,6 @@
 /* eslint-disable jsdoc/require-jsdoc */
 import { ExtractResolvedPayloadActions, resolveSimpleAction } from "$actions";
-import { hasId, rejectItemWithId, updateItemWithId } from "$root/utils";
+import { rejectItemWithId, updateItemWithId } from "$root/utils";
 import { CharacterRemovalStrategy, FullSheet } from "$sheets/types";
 import { itemIsCarriedByCharacterId } from "$sheets/utils";
 import {
@@ -23,7 +23,7 @@ import {
 	ResolvedInventoryStoreAction,
 } from "$sheets/store/inventoryActions";
 
-type CharacterDialogStateProps =
+export type CharacterDialogStateProps =
 	| {
 			mode: "closed";
 	  }
@@ -38,12 +38,14 @@ type CharacterDialogStateProps =
 			mode: "new-character";
 	  };
 
-type InventoryStoreProps = {
+export type InventoryStoreProps = {
 	sheet: FullSheet;
 	resolvedActionIds: string[];
 	ui: {
 		characterDialog: CharacterDialogStateProps;
 		sheetNameDialogIsOpen: boolean;
+		carriedByFilter: string[];
+		categoryFilter: string[];
 	};
 };
 
@@ -61,66 +63,10 @@ const initialInventoryStoreState: InventoryStoreProps = {
 			mode: "closed",
 		},
 		sheetNameDialogIsOpen: false,
+		carriedByFilter: [],
+		categoryFilter: [],
 	},
 };
-
-/* #region  SELECTORS */
-/**
- * Utility for defining a selector that pulls from the sheet field
- *
- * @param selector the selector
- */
-export const fromSheet = <Selection>(selector: (p: FullSheet) => Selection) => (
-	state: InventoryStoreProps
-) => selector(state.sheet);
-
-/* eslint-disable jsdoc/require-jsdoc */
-type InventoryStoreSelector<Selection> = (
-	state: InventoryStoreProps
-) => Selection;
-
-export const selectCharacterDialogMode: InventoryStoreSelector<
-	CharacterDialogStateProps["mode"]
-> = (s) => s.ui.characterDialog.mode;
-
-export const selectCharacterBeingEdited: InventoryStoreSelector<Character | null> = (
-	s
-) => {
-	const idOfCharacterBeingEdited = match(s.ui.characterDialog)
-		.with({ mode: "edit" }, ({ data }) => data.characterId)
-		.otherwise(() => null);
-
-	if (!idOfCharacterBeingEdited) return null;
-
-	return s.sheet.characters.find(hasId(idOfCharacterBeingEdited)) ?? null;
-};
-
-export const selectCharacterCarriedItems = (
-	characterId: string
-): InventoryStoreSelector<Item[]> =>
-	fromSheet((sheet) =>
-		sheet.items.filter((item) => item.carriedByCharacterId === characterId)
-	);
-
-export const selectItemsCarriedByCharacterBeingEdited: InventoryStoreSelector<
-	Item[]
-> = (state) => {
-	const characterBeingEdited = selectCharacterBeingEdited(state);
-	const items = selectCharacterCarriedItems(characterBeingEdited?.id ?? "")(
-		state
-	);
-	return items;
-};
-
-export const selectCharacters: InventoryStoreSelector<Character[]> = fromSheet(
-	(sheet) => sheet.characters
-);
-
-export const selectSheetName: InventoryStoreSelector<string> = fromSheet(
-	(s) => s.name
-);
-/* eslint-enable jsdoc/require-jsdoc */
-/* #endregion */
 
 export const useLastInventoryStoreAction = createState(() => ({
 	lastAction: null as FinalInventoryStoreAction | null,
